Refactoring changes in Microsoft.AspNet.PiplelineCore
* Moving multiple partial classes in a single cs file into multiple files
* Moving types to Internal namespace \ directory and making them public.
* Renaming ICanHasQuery to IQueryCollectionProvider
