+++ src/Microsoft.AspNet.PipelineCore/Collections/HeaderDictionary.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNet.Abstractions.Infrastructure;
 using Microsoft.AspNet.Abstractions;
-using Microsoft.AspNet.PipelineCore.Infrastructure;
+using Microsoft.AspNet.PipelineCore.Internal;
 
 namespace Microsoft.AspNet.PipelineCore.Collections
 {
