+++ src/Microsoft.AspNet.PipelineCore/Collections/ResponseCookieCollection.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using Microsoft.AspNet.Abstractions.Infrastructure;
 using Microsoft.AspNet.Abstractions;
+using Microsoft.AspNet.Abstractions.Infrastructure;
 
 namespace Microsoft.AspNet.PipelineCore.Collections
 {
