--- src/Microsoft.AspNet.PipelineCore/ICanHasQuery.cs
@@ -1,9 +1,0 @@
-using Microsoft.AspNet.Abstractions;
-
-namespace Microsoft.AspNet.PipelineCore
-{
-    public interface ICanHasQuery
-    {
-        IReadableStringCollection Query { get; }
-    }
-}
