+++ src/Microsoft.AspNet.PipelineCore/Internal/Constants.cs
